+++ tests/backend/test_session.py
@@ -10,6 +10,7 @@ from pathlib import Path
 
 import lintrans
 from lintrans.gui.session import Session
+from lintrans.gui.settings import DisplaySettings
 from lintrans.matrices.wrapper import MatrixWrapper
 
 from conftest import get_test_wrapper
@@ -18,7 +19,7 @@ from conftest import get_test_wrapper
 def test_save_and_load(tmp_path: Path, test_wrapper: MatrixWrapper) -> None:
     """Test that sessions save and load and return the same matrix wrapper."""
     points = [(1, 0), (-2, 3), (3.2, -10), (0, 0), (-2, -3), (2, -1.3)]
-    session = Session(matrix_wrapper=test_wrapper, polygon_points=points)
+    session = Session(matrix_wrapper=test_wrapper, polygon_points=points, display_settings=DisplaySettings())
 
     path = str((tmp_path / 'test.lt').absolute())
     session.save_to_file(path)
@@ -26,6 +27,7 @@ def test_save_and_load(tmp_path: Path, test_wrapper: MatrixWrapper) -> None:
     loaded_session, version, extra_attrs = Session.load_from_file(path)
     assert loaded_session.matrix_wrapper == get_test_wrapper()
     assert loaded_session.polygon_points == points
+    assert loaded_session.display_settings == DisplaySettings()
 
     assert version == lintrans.__version__
     assert not extra_attrs