+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Input and output vectors can now be shown or hidden independently
 - Installers for Windows and Linux
 - The option to let lintrans update itself in the background
+- Display settings are now saved as part of the session file
 
 ### Fixed